@@ -2,11 +2,13 @@
   "name": "dashboard-api",
   "version": "1.0.0",
   "main": "main.js",
-  "type": "commonjs",
   "scripts": {
     "start": "node ./dist/main.js",
+    "dev": "nodemon",
+    "dev:inspect":"nodemon -e ts,json --exec node --inspect=localhost:9222 -r ts-node/register src/main.ts",
     "build": "tsc",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "eslint": "eslint ./src/**",
+    "eslint:fix": "eslint ./src/** --fix"
   },
   "author": "",
   "license": "ISC",
@@ -24,7 +26,9 @@
     "eslint": "^8.0.1",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^4.0.0",
+    "nodemon": "^3.1.11",
     "prettier": "^2.4.1",
-    "typescript": "^4.4.4"
+    "ts-node": "^10.9.2",
+    "typescript": "^5.9.3"
   }
 }
