@@ -1,41 +1,38 @@
-import { Response, Router } from "express";
-import { Route } from "./route.interface";
-import { LoggerType } from "../logger/logger.interface";
-import { injectable } from "inversify";
-import 'reflect-metadata'
+import { Response, Router } from 'express';
+import { Route } from './route.interface';
+import { LoggerType } from '../logger/logger.interface';
+import { injectable } from 'inversify';
+import 'reflect-metadata';
 
 @injectable()
 export abstract class BaseController {
-    private readonly _router: Router
+	private readonly _router: Router;
 
-    constructor(private logger: LoggerType) {
-        this._router = Router()
-    }
+	constructor(private logger: LoggerType) {
+		this._router = Router();
+	}
 
-    get router() {
-        return this._router
-    }
+	get router(): Router {
+		return this._router;
+	}
 
-    public send<T>(res: Response, code: number,  message: T) {
-        res.type('application/json')
-        return res.status(code).json(message)
-    }
+	public send<T>(res: Response, code: number, message: T): Response<any, Record<string, any>> {
+		res.type('application/json');
+		return res.status(code).json(message);
+	}
 
-    public ok<T>(res: Response,  message: T) {
-        return this.send<T>(res, 200, message)
-    }
+	public ok<T>(res: Response, message: T): Response<any, Record<string, any>> {
+		return this.send<T>(res, 200, message);
+	}
 
-    public created(res: Response) {
-        return res.sendStatus(201)
-    }
+	public created(res: Response): Response<any, Record<string, any>> {
+		return res.sendStatus(201);
+	}
 
-    protected bindRouter(routes: Route[]) {
-        for(const route of routes) {
-            this.logger.log(`[${route.method}] ${route.path}`)
-            this.router[route.method](route.path, route.func.bind(this))
-        }
-    }
-
-
-
+	protected bindRouter(routes: Route[]): void {
+		for (const route of routes) {
+			this.logger.log(`[${route.method}] ${route.path}`);
+			this.router[route.method](route.path, route.func.bind(this));
+		}
+	}
 }
