@@ -1,5 +1,5 @@
-import { NextFunction, Request, Response } from "express";
+import { NextFunction, Request, Response } from 'express';
 
 export interface ExeptionFilter {
-    catch:(err: Error, req: Request, res: Response, next: NextFunction) => void
+	catch: (err: Error, req: Request, res: Response, next: NextFunction) => void;
 }
