@@ -1,7 +1,6 @@
 export interface LoggerType {
-    logger: unknown;
-    log:(...args: unknown[]) => void
-    error:(...args: unknown[]) => void
-    warn:(...args: unknown[]) => void
-    
+	logger: unknown;
+	log: (...args: unknown[]) => void;
+	error: (...args: unknown[]) => void;
+	warn: (...args: unknown[]) => void;
 }
