@@ -1,32 +1,34 @@
-import { Container, ContainerModule } from "inversify";
-import { App } from "./app";
-import { ExeptionFilters } from "./errors/exeption.filter";
-import { LoggerService } from "./logger/logger.service";
-import { UserController } from "./users/user.controller";
-import { LoggerType } from "./logger/logger.interface";
-import { TYPES } from "./types";
-import { ExeptionFilter } from "./errors/exeption.filter.interface";
-import { UserControllerType } from "./users/user.controller.interface";
+import { Container, ContainerModule } from 'inversify';
+import { App } from './app';
+import { ExeptionFilters } from './errors/exeption.filter';
+import { LoggerService } from './logger/logger.service';
+import { UserController } from './users/user.controller';
+import { LoggerType } from './logger/logger.interface';
+import { TYPES } from './types';
+import { ExeptionFilter } from './errors/exeption.filter.interface';
+import { UserControllerType } from './users/user.controller.interface';
 
-
-export const appBindings = new ContainerModule((bind) => {
-    bind.bind<LoggerType>(TYPES.LoggerType).to(LoggerService)
-    bind.bind<ExeptionFilter>(TYPES.ExeptionFilter).to(ExeptionFilters)
-    bind.bind<UserControllerType>(TYPES.UserController).to(UserController)
-    bind.bind<App>(TYPES.Application).to(App)   
-})
-
-
-function bootstrap() {
-    const appContainer = new Container()
-    appContainer.load(appBindings)
-    const app = appContainer.get<App>(TYPES.Application)
-    app.init()
-    return { app, appContainer }
+export interface BootstrapReturn {
+	app: App;
+	appContainer: Container;
 }
 
+export const appBindings = new ContainerModule((bind) => {
+	bind.bind<LoggerType>(TYPES.LoggerType).to(LoggerService);
+	bind.bind<ExeptionFilter>(TYPES.ExeptionFilter).to(ExeptionFilters);
+	bind.bind<UserControllerType>(TYPES.UserController).to(UserController);
+	bind.bind<App>(TYPES.Application).to(App);
+});
 
-export const { app, appContainer } = bootstrap()
+function bootstrap(): BootstrapReturn {
+	const appContainer = new Container();
+	appContainer.load(appBindings);
+	const app = appContainer.get<App>(TYPES.Application);
+	app.init();
+	return { app, appContainer };
+}
+
+export const { app, appContainer } = bootstrap();
 
 /* 
 import express from 'express'
